A number of changes related to the server and relative paths
This is somewhat of an intermediate state; see below about cli-module-loader.

There are also lots of pathing issues that can come up here; a next step is to
write a simple repository that npm installs pyret with -g, and tests
compilation commands from various working directories. This will ensure that
local testing doesn't have random state floating around in my directories that
make this work when it shouldn't.

- Make much more use of the `this-pyret-dir` parameter:
  - Allow for `$PYRET` in config files to mean substitute with
    `this-pyret-dir`. This allows us to consolidate all the config.json files
    into one.
  - Remove lots of defaults from `pyret.js` in favor of calculating the path to
    the appropriate JS file for e.g. handalone, etc
  - Copy more files into the `build/` directory for the appropriate phase in
    order to centralize the location of dependencies. For example, jglr files
    and json config files are now copied into build/phaseX/js
  - Use the location of the pyret.js script to find the compiler (currently
    this looks for phaseA). Since this is the main entry-point, it has to
    somehow hard-code a default compiler to look for.

- Run the compiler in async mode, and use `schedulePause` along with SIGINT and
  a shutdown message to control the server.  The client-lib still has logic for
  killing the server via a PID file, but this logic should eventually be
  removed. The server should create and manage the socket file, and there
  should be an option like --clean that causes the server to unlink any old
  sockets it finds when it starts up. The client should just try to connect to
  the well-known socket for a few seconds rather than looking for a pid file,
  and start the server if the appropriate socket file doesn't exist.

- Send SIGINT (Ctrl-C) events from the client to the server via a special
  "stop" command, which breaks the current Pyret execution but keeps the server
  running.

- Add a new compiler option for a list of directories in which to _look_ for
  pre-built builtin modules like lists and ffi. These should have half of the
  semantics of the writable cache directory that defaults to compiled/ – the
  cli-module-loader will look there for cached versions of files, but only
  write to the specified destination. Essentially all the new logic in
  cli-module-loader is handling this new option. This has introduced a stupid
  slowdown in the cache-checking logic when the compiler is used from the
  server that I need to track down. I believe many .arr files are getting
  parsed more than they need to when their compiled versions already exist. All
  of this caching logic could use a long look anyway, so I'll be poking into
  this more.

- Change the compiler to resolve the name of the output .jarr file relative to
  the provided base-dir, rather than relative to the current working directory.

- Add various intelligent choices revolving around the current working
  directory to the options that the client sends to the server.

+++ src/server/client-lib.js
@@ -10,7 +10,7 @@ const LOG = 3;
 const WARN = 2;
 const ERROR = 1;
 const SILENT = 0;
-var LOG_LEVEL = WARN;
+var LOG_LEVEL = INFO;
 
 function makeLogger(level) {
   return function(...args) {
@@ -91,25 +91,39 @@ function start(options) {
     const compileTarget = path.join(localParleyDir, "compiled");
     options["pyret-options"]["compiled-dir"] = compileTarget;
   }
+  if(!options["pyret-options"]["base-dir"]) {
+    const baseDir = path.resolve(path.join(localParleyDir, ".."));
+    options["pyret-options"]["base-dir"] = baseDir;
+  }
 
   function shutdown() {
     try {
-      const pid = Number(fs.readFileSync(pidFile));
-      process.kill(pid, 'SIGINT');
-      fs.unlinkSync(pidFile);
-      fs.unlinkSync(portFile);
-      info("Sent kill signal to " + pid);
+      const client = new WebSocket("ws+unix://" + portFile);
+      client.on('error', function(err) {
+        error('Connection error: ' + err.toString() + ".");
+        error('You can try again, and the Pyret server may restart. If you see this error repeatedly, report it as a bug.');
+      });
+      client.on('open', function(connection) {
+        client.send(JSON.stringify({ command: 'shutdown' }));
+      });
+      
+      tryToRemoveFiles();
     }
     catch(e) {
-      info("No process to quit: " + e);
-      fs.unlinkSync(pidFile);
-      fs.unlinkSync(portFile);
+      info("Error during shutdown: " + e);
+      tryToRemoveFiles();
     }
+  }
+
+  function tryToRemoveFiles() {
+    if(fs.existsSync(pidFile)) { fs.unlinkSync(pidFile); }
+    if(fs.existsSync(portFile)) { fs.unlinkSync(portFile); }
   }
 
   if(options.client.shutdown) {
     shutdown();
-    process.exit(0);
+    return;
+    //process.exit(0);
   }
 
   function makeSocketAndConnect() {
@@ -120,12 +134,22 @@ function start(options) {
       shutdown();
     });
 
+    function sigint() {
+      log("Caught interrupt signal during compile job, stopping compile job");
+      client.send(JSON.stringify({ command: 'stop' }));
+      process.exit(0);
+    }
+
     client.on('close', function() {
       info('parley connection closed');
+      process.removeListener('SIGINT', sigint);
     });
      
     client.on('open', function(connection) {
       info('parley protocol connected');
+
+      process.on('SIGINT', sigint);
+
       client.on('message', function(message) {
         const parsed = JSON.parse(message);
         if(parsed.type === 'echo-log') {
@@ -143,13 +167,16 @@ function start(options) {
           process.stderr.write(parsed.contents);
         }
       });
-      client.send(JSON.stringify(options['pyret-options']));
+
+
+      var forMessage = { command: "compile", compileOptions: JSON.stringify(options['pyret-options']) };
+      client.send(JSON.stringify(forMessage));
     });
     return client;
   }
 
 
-  function startupServer(port) {
+  function startupServer(port, wait) {
     const child = childProcess.fork(
       serverModule,
       ["-serve", "--port", port],
@@ -159,19 +186,25 @@ function start(options) {
       } // To send messages on completion of startup
     );
 
-    return new Promise((resolve, reject) => {
-      child.on('message', function(msg) {
-        if(msg.type === 'success') {
-          child.unref();
-          child.disconnect();
-          fs.writeFileSync(pidFile, String(child.pid));
-          resolve(msg);
-        }
-        else {
-          reject(msg);
-        }
+    if(wait) {
+      return new Promise((resolve, reject) => {
+        child.on('message', function(msg) {
+          if(msg.type === 'success') {
+            child.unref();
+            child.disconnect();
+            resolve(msg);
+          }
+          else {
+            reject(msg);
+          }
+        });
       });
-    });
+    }
+    else {
+      fs.writeFileSync(pidFile, String(child.pid));
+      child.unref();
+      child.disconnect();
+    }
   }
 
   const connectURL = portFile;
@@ -182,7 +215,7 @@ function start(options) {
     if(pidStats.mtime.getTime() < compilerStats.mtime.getTime()) {
       info("A running server was found, but the chosen compiler (" + serverModule + ") is newer than the server.  Restarting...");
       shutdown();
-      startupServer(portFile)
+      startupServer(portFile, true)
         .then(() => makeSocketAndConnect())
         .catch((err) => error('Starting up the server failed: ', err));
     }
@@ -192,16 +225,16 @@ function start(options) {
     }
   // Otherwise, try starting up a server and waiting for it, then doing the work
   } else {
-    info("No pidFile found, starting up server");
-    startupServer(portFile)
-      .then(() => makeSocketAndConnect())
-      .catch((err) => console.error('Starting up the server failed: ', err));
+    if (fs.existsSync(portFile)) {
+      makeSocketAndConnect();
+    }
+    else {
+      info("No pid file or socket found, starting up server");
+      startupServer(portFile, true)
+        .then(() => makeSocketAndConnect())
+        .catch((err) => console.error('Starting up the server failed: ', err));
+    }
   }
-
-  process.on('SIGINT', function() {
-    info("Caught interrupt signal, shutting down server");
-    shutdown();
-  });
 
 }
